Refactor: split db.go
